test: centralize Tauri e2e setup
