+++ tests/app.e2e.test.js
@@ -1,51 +1,9 @@
-import { spawn, spawnSync } from 'child_process';
-import path from 'path';
-import { waitTauriDriverReady } from '@crabnebula/tauri-driver';
-import { beforeAll, afterAll, test, expect } from 'vitest';
-import { remote } from 'webdriverio';
-
-let browser;
-let tauriDriver;
+import { test, expect } from 'vitest';
 
 const parseXp = (text) => {
   const match = /XP: (\d+)/.exec(text);
   return match ? Number(match[1]) : 0;
 };
-
-beforeAll(async () => {
-  const tauriDir = path.resolve(__dirname, '../src-tauri');
-  spawnSync('npx', ['tauri', 'build', '--debug'], {
-    cwd: tauriDir,
-    stdio: 'inherit',
-  });
-
-  tauriDriver = spawn('npx', ['tauri-driver'], {
-    stdio: 'inherit',
-    shell: true,
-  });
-  await waitTauriDriverReady();
-
-  const appPath = path.resolve(tauriDir, 'target/debug/zimbo-panel');
-  browser = await remote({
-    hostname: '127.0.0.1',
-    port: 4444,
-    capabilities: {
-      'tauri:options': {
-        application: appPath,
-      },
-    },
-    logLevel: 'error',
-  });
-});
-
-afterAll(async () => {
-  if (browser) {
-    await browser.deleteSession();
-  }
-  if (tauriDriver) {
-    tauriDriver.kill();
-  }
-});
 
 test('character flow with save and load', async () => {
   // wait for XP text to be available
