+++ vitest.e2e.config.js
@@ -4,5 +4,9 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     include: ['tests/**/*.e2e.test.js', 'tests/**/*.e2e.spec.ts'],
+    setupFiles: ['tests/e2e.setup.ts'],
+    threads: {
+      singleThread: true,
+    },
   },
 });
